exporter/requirements.txt
beautifulsoup4==4.12.3
markdownify==0.11.6
Pillow==10.2.0
python-dateutil==2.8.2